python-analyzer/requirements.txt
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Events==0.5
fastapi==0.120.0
fuzzywuzzy==0.18.0
h11==0.16.0
httptools==0.7.1
idna==3.11
opensearch-py==3.0.0
packaging==25.0
pillow==12.0.0
pydantic==2.12.3
pydantic_core==2.41.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
pyzbar==0.1.9
RapidFuzz==3.14.1
regex==2025.10.23
requests==2.32.5
selectolax==0.4.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1; sys_platform != "win32"  # uvloop doesn't support Windows
watchfiles==1.1.1
websockets==15.0.1
httpx
pandas
sqlalchemy
psycopg2-binary

minio
spacy
redis
python-dotenv
